bugfix : problème de la mise à jour du produit a été réglé.

+++ Back/app/Http/Controllers/AdminProductController.php
@@ -93,18 +93,25 @@ class AdminProductController extends Controller
                 'category_id' => 'exists:categories,id',
             ]);
 
+            $product->update($validated);
+
             if ($request->hasFile('image')) {
-                if ($product->image) {
+                if ($product->image && $product->image->name    ) {
                     Storage::disk('public')->delete($product->image->name);
                     Log::info('Old image deleted', ['path' => $product->image->name]);
                 }
-                $validated['image'] = $request->file('image')->store('product_images', 'public');
+                    $validated['image'] = $request->file('image')->store('product_images', 'public');
                 Log::info('New image stored', ['path' => $validated['image']]);
+
+                if ($product->image) {
+                    $product->image()->update(['name'=> $validated['image']]);
+                    Log::info('Product updated', ['id' => $product->id]);
+                } else {
+                    $product->image()->create(['name'=> $validated['image']]);
+
+                }
             }
 
-            $product->update($validated);
-            $product->image()->update(['name'=> $validated['image']]);
-            Log::info('Product updated', ['id' => $product->id]);
             return response()->json($product);
         } catch (\Exception $e) {
             Log::error('Failed to update product', [
